lnnode: split daemon to two files
In addition to splitting the daemon package the Start and Stop function
are pareameterless and returns error.
